+++ NVEncCore/rgy_prm.h
@@ -1347,7 +1347,7 @@ struct AudioSelect {
     tstring  encCodecProfile; //音声エンコードのコーデックのプロファイル
     int      encBitrate;      //音声エンコードに選択した音声トラックのビットレート
     int      encSamplingRate;      //サンプリング周波数
-    int      addDelayMs;           //追加する音声の遅延(millisecond)
+    double   addDelayMs;           //追加する音声の遅延(millisecond)
     tstring  extractFilename;      //抽出する音声のファイル名のリスト
     tstring  extractFormat;        //抽出する音声ファイルのフォーマット
     tstring  filter;               //音声フィルタ
